Break apart file into smaller modules for code clarity
