+++ src/middleware/auth/AuthConfig.ts
@@ -1,0 +1,4 @@
+export interface AuthConfig {
+  apiKey?: string
+  secretOrKey?: string
+}
